+++ Sources/StreamChat/Models/Channel.swift
@@ -29,6 +29,9 @@ public struct ChatChannel {
     /// If the channel was deleted, this field contains the date of the deletion.
     public let deletedAt: Date?
     
+    /// Flag for representing hidden state for the channel.
+    public let isHidden: Bool
+    
     /// The user which created the channel.
     public let createdBy: ChatUser?
     
@@ -153,6 +156,7 @@ public struct ChatChannel {
         createdAt: Date = .init(),
         updatedAt: Date = .init(),
         deletedAt: Date? = nil,
+        isHidden: Bool,
         createdBy: ChatUser? = nil,
         config: ChannelConfig = .init(),
         isFrozen: Bool = false,
@@ -179,6 +183,7 @@ public struct ChatChannel {
         self.createdAt = createdAt
         self.updatedAt = updatedAt
         self.deletedAt = deletedAt
+        self.isHidden = isHidden
         self.createdBy = createdBy
         self.config = config
         self.isFrozen = isFrozen