+++ Sources/StreamChat/WebSocketClient/EventMiddlewares/ChannelVisibilityEventMiddleware.swift
@@ -14,18 +14,27 @@ struct ChannelVisibilityEventMiddleware: EventMiddleware {
                     throw ClientError.ChannelDoesNotExist(cid: event.cid)
                 }
                 
-                channelDTO.hiddenAt = nil
+                channelDTO.isHidden = false
                 
             case let event as ChannelHiddenEventDTO:
                 guard let channelDTO = session.channel(cid: event.cid) else {
                     throw ClientError.ChannelDoesNotExist(cid: event.cid)
                 }
                 
-                channelDTO.hiddenAt = event.createdAt
+                channelDTO.isHidden = true
                 
                 if event.isHistoryCleared {
                     channelDTO.truncatedAt = event.createdAt
                 }
+                
+            // New Message will unhide the channel
+            // but we won't get `ChannelVisibleEvent` for this case
+            case let event as MessageNewEventDTO:
+                guard let channelDTO = session.channel(cid: event.cid) else {
+                    throw ClientError.ChannelDoesNotExist(cid: event.cid)
+                }
+                
+                channelDTO.isHidden = false
                 
             default:
                 break